+++ handlers/handler_subject.go
@@ -66,7 +66,7 @@ func (handler SubjectHandler) Create(c *gin.Context) {
 			existingSubject := m.Subject{}
 			existingSubjectQuery := handler.db.Where("id != ? AND subject_code = ? AND created_by = ? AND deleted_at is NULL", newSubject.Id, c.PostForm("subject_code"), GetCreator(c)).First(&existingSubject)
 			
-			if (c.PostForm("for_deletion") == "") {
+			if (c.PostForm("for_deletion") == "false") {
 				if existingSubjectQuery.RowsAffected > 0 {
 					respond(http.StatusBadRequest, "Subject code already existing.", c, true)
 				} else {
@@ -82,11 +82,15 @@ func (handler SubjectHandler) Create(c *gin.Context) {
 					}
 				}
 			} else {
-				delete := handler.db.Delete(&existingSubjectById)
-				if delete.RowsAffected > 0 {
-					respond(http.StatusOK, "Record successfully deleted.", c, false)
+				if (c.PostForm("for_deletion") == "true") {
+					delete := handler.db.Delete(&existingSubjectById)
+					if delete.RowsAffected > 0 {
+						respond(http.StatusOK, "Record successfully deleted.", c, false)
+					} else {
+						respond(http.StatusBadRequest, delete.Error.Error(), c, true)
+					}
 				} else {
-					respond(http.StatusBadRequest, delete.Error.Error(), c, true)
+					respond(http.StatusBadRequest, "Invalid action.", c, true)
 				}
 			}	
 		} else {
